+++ src/LinkMobility/LinkMobilityResponse.cs
@@ -1,20 +1,38 @@
 ﻿namespace AMWD.Net.Api.LinkMobility
 {
+	/// <summary>
+	/// A generic response of the LinkMobility API.
+	/// </summary>
 	public class LinkMobilityResponse
 	{
+		/// <summary>
+		/// Contains the message id defined in the request.
+		/// </summary>
 		[JsonProperty("clientMessageId")]
-		public string ClientMessageId { get; set; }
+		public string? ClientMessageId { get; set; }
 
+		/// <summary>
+		/// The actual number of generated SMS.
+		/// </summary>
 		[JsonProperty("smsCount")]
-		public int SmsCount { get; set; }
+		public int? SmsCount { get; set; }
 
+		/// <summary>
+		/// Status code
+		/// </summary>
 		[JsonProperty("statusCode")]
-		public StatusCodes StatusCode { get; set; }
+		public StatusCodes? StatusCode { get; set; }
 
+		/// <summary>
+		/// Description of the response status code.
+		/// </summary>
 		[JsonProperty("statusMessage")]
-		public string StatusMessage { get; set; }
+		public string? StatusMessage { get; set; }
 
+		/// <summary>
+		/// Unique identifier that is set after successful processing of the request.
+		/// </summary>
 		[JsonProperty("transferId")]
-		public string TransferId { get; set; }
+		public string? TransferId { get; set; }
 	}
 }